requirements.txt
minio==7.1.2
python-dotenv==0.19.2
tensorboard==2.7.0
tensorflow==2.7.0
tensorflow_hub==0.12.0
tensorflow_io==0.23.1
